refactor: extract AdvisorInventoryTestMixin to shared test module
Move replicate_to_advisor_inventory() and the test mixin class from
test_system_views.py to tests/__init__.py so all test modules can
reuse the same flag-on test infrastructure.
